requirements.txt
ale-py==0.10.1
box2d-py==2.3.5
flax==0.7.2
gymnasium==1.0.0
jaraco.collections==5.1.0
matplotlib==3.7.5
moviepy==1.0.3
mujoco==3.2.3
mujoco-py==2.1.2.14
opencv-python==4.11.0.86
pip-chill==1.0.3
pipreqs==0.5.0
pygame==2.6.1
swig==4.3.0
tinycss2==1.2.1
tomli==2.0.1
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
